+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-- More daily quests checked
-- Miner quests added
-- Add sponsor button to quest window with hide config option
+- New IPC: RedoLookupIndex resolves quest ID to chapter name and index in quest list
+- DT msq part 1 checked for bug report